Improve evaluate_at
* Split out part of the code
* Return indices such that they can be used directly for isel (this
breaks previous behaviour, but is much nicer to use)
* Should now work with additional dimensions
